backend/requirements.txt
streamlit
pyautogen
Django
djangorestframework
python-dotenv
django-cors-headers